add db function for put, quiz api call
